+++ apps/ledger-live-desktop/tests/fixtures/common.ts
@@ -107,6 +107,8 @@ export const test = base.extend<TestFixtures>({
   page: async ({ electronApp }, use, testInfo) => {
     // app is ready
     const page = await electronApp.firstWindow();
+    // we need to give enough time for the playwright app to start. when the CI is slow, 30s was apprently not enough.
+    page.setDefaultTimeout(99000);
 
     // record all logs into an artifact
     const logFile = testInfo.outputPath("logs.log");