+++ apps/ledger-live-mobile/.eslintrc.js
@@ -44,8 +44,8 @@ module.exports = {
       {
         patterns: [
           {
-            group: ["@ledgerhq/live-common/lib/*"],
-            message: '🚨 Please when importing from live-common, remove the "/lib/" in the path 🚨',
+            group: ["@ledgerhq/live-common/lib/**", "@ledgerhq/live-common/lib-es/**"],
+            message: "Please remove the /lib import from live-common import.",
           },
         ],
       },