+++ libs/coin-evm/src/__tests__/unit/preload.unit.test.ts
@@ -3,7 +3,7 @@ jest.useFakeTimers();
 
 import network from "@ledgerhq/live-network/network";
 import evms from "@ledgerhq/cryptoassets/data/evm/index";
-import { ERC20Token } from "@ledgerhq/cryptoassets/types";
+import { BEP20Token, ERC20Token } from "@ledgerhq/cryptoassets/types";
 import * as CALTokensAPI from "@ledgerhq/cryptoassets/tokens";
 import { getCryptoCurrencyById } from "@ledgerhq/cryptoassets";
 import { fetchERC20Tokens, hydrate, preload } from "../../preload";
@@ -34,15 +34,29 @@ const usdtDefinition: ERC20Token = [
   undefined,
   undefined,
 ];
+const binanceDaiDefinition: BEP20Token = [
+  "bsc",
+  "binance-peg_dai_token",
+  "DAI",
+  18,
+  "Binance-Peg Dai Token",
+  "3044022032f0a880722af8c9e2196b5c0fc5273e2088f23692bdd2b35f6cf41c4001213f02205226e2023e409c73b031c790c64ae24db67c04b0aefd0d979b8c5002ca969b7b",
+  "0x1AF3F329e8BE154074D8769D1FFa4eE058B1DBc3",
+  true,
+  false,
+];
 const currency1 = getCryptoCurrencyById("ethereum"); // chain id 1
+const currency2 = getCryptoCurrencyById("bsc"); // chain id 56
 
 jest.mock("@ledgerhq/live-network/network");
 jest.mock("@ledgerhq/cryptoassets/data/evm/index", () => ({
   get tokens(): {
     1: ERC20Token[];
+    56: BEP20Token[];
   } {
     return {
       1: [usdcDefinition],
+      56: [binanceDaiDefinition],
     };
   },
 }));
@@ -72,12 +86,20 @@ describe("EVM Family", () => {
         expect(network).toHaveBeenCalledTimes(2);
       });
 
-      it("should load dynamically the tokens", async () => {
+      it("should load dynamically the ERC20 tokens", async () => {
         // @ts-expect-error not casted as jest mock
         network.mockResolvedValue({ data: [usdtDefinition] });
 
         const tokens = await fetchERC20Tokens(currency1);
         expect(tokens).toEqual([usdtDefinition]);
+      });
+
+      it("should load dynamically the BEP20 tokens", async () => {
+        // @ts-expect-error not casted as jest mock
+        network.mockResolvedValue({ data: [binanceDaiDefinition] });
+
+        const tokens = await fetchERC20Tokens(currency2);
+        expect(tokens).toEqual([binanceDaiDefinition]);
       });
 
       it("should fallback on local CAL on dynamic CAL error", async () => {
@@ -88,23 +110,29 @@ describe("EVM Family", () => {
 
         const tokens = await fetchERC20Tokens(currency1);
         expect(tokens).toEqual([usdcDefinition]);
+        const bep20tokens = await fetchERC20Tokens(currency2);
+        expect(bep20tokens).toEqual([binanceDaiDefinition]);
       });
 
       it("should load erc20 tokens from local CAL when dynamic CAL undefined", async () => {
         const tokens = await fetchERC20Tokens(currency1);
         expect(tokens).toEqual([usdcDefinition]);
+        const bep20tokens = await fetchERC20Tokens(currency2);
+        expect(bep20tokens).toEqual([binanceDaiDefinition]);
       });
 
       it("should load erc20 tokens from local CAL when dynamic CAL is empty []", async () => {
         // @ts-expect-error not casted as jest mock
         network.mockResolvedValue({
           data: {
-            tokens: { 1: [] },
+            tokens: { 1: [], 56: [] },
           },
         });
 
-        const tokens = await fetchERC20Tokens(currency1);
-        expect(tokens).toEqual([usdcDefinition]);
+        const erc20tokens = await fetchERC20Tokens(currency1);
+        expect(erc20tokens).toEqual([usdcDefinition]);
+        const bep20tokens = await fetchERC20Tokens(currency2);
+        expect(bep20tokens).toEqual([binanceDaiDefinition]);
       });
 
       it("should return empty [] if dynamic CAL fails and local CAL fails", async () => {
@@ -126,6 +154,17 @@ describe("EVM Family", () => {
           CALTokensAPI.convertERC20(usdcDefinition),
         ]);
       });
+
+      it("should register BEP20 tokens", async () => {
+        jest.spyOn(CALTokensAPI, "addTokens").mockImplementationOnce(() => null);
+
+        const tokens = await preload(currency2);
+
+        expect(tokens).toEqual([binanceDaiDefinition]);
+        expect(CALTokensAPI.addTokens).toHaveBeenCalledWith([
+          CALTokensAPI.convertBEP20(binanceDaiDefinition),
+        ]);
+      });
     });
 
     describe("hydrate", () => {
@@ -134,16 +173,27 @@ describe("EVM Family", () => {
       });
 
       it("should return void", () => {
-        expect(hydrate(undefined)).toBe(undefined);
+        expect(hydrate(undefined, currency1)).toBe(undefined);
+        expect(hydrate(undefined, currency2)).toBe(undefined);
       });
 
-      it("should register tokens", async () => {
+      it("should register ERC20 tokens", async () => {
         jest.spyOn(CALTokensAPI, "addTokens").mockImplementationOnce(() => null);
 
-        await hydrate([usdcDefinition]);
+        await hydrate([usdcDefinition], currency1);
 
         expect(CALTokensAPI.addTokens).toHaveBeenCalledWith([
           CALTokensAPI.convertERC20(usdcDefinition),
+        ]);
+      });
+
+      it("should register BEP20 tokens", async () => {
+        jest.spyOn(CALTokensAPI, "addTokens").mockImplementationOnce(() => null);
+
+        await hydrate([binanceDaiDefinition], currency2);
+
+        expect(CALTokensAPI.addTokens).toHaveBeenCalledWith([
+          CALTokensAPI.convertBEP20(binanceDaiDefinition),
         ]);
       });
     });