+++ libs/ledgerjs/packages/hw-app-eth/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ledgerhq/hw-app-eth
 
+## 6.34.8
+
+### Patch Changes
+
+- Updated dependencies [[`ce18546c0a`](https://github.com/LedgerHQ/ledger-live/commit/ce18546c0a0b9dd5ed78b1745cac19b7eef7b5eb), [`fbeebfe04b`](https://github.com/LedgerHQ/ledger-live/commit/fbeebfe04b297b33ec258440b694cdfb6213af24)]:
+  - @ledgerhq/types-live@6.41.1
+  - @ledgerhq/cryptoassets@11.0.1
+  - @ledgerhq/domain-service@1.1.13
+  - @ledgerhq/evm-tools@1.0.9
+
+## 6.34.8-hotfix.1
+
+### Patch Changes
+
+- Updated dependencies [[`fbeebfe04b`](https://github.com/LedgerHQ/ledger-live/commit/fbeebfe04b297b33ec258440b694cdfb6213af24)]:
+  - @ledgerhq/cryptoassets@11.0.1-hotfix.0
+  - @ledgerhq/domain-service@1.1.13-hotfix.1
+  - @ledgerhq/evm-tools@1.0.9-hotfix.0
+
+## 6.34.8-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies [[`ce18546c0a`](https://github.com/LedgerHQ/ledger-live/commit/ce18546c0a0b9dd5ed78b1745cac19b7eef7b5eb)]:
+  - @ledgerhq/types-live@6.41.1-hotfix.0
+  - @ledgerhq/domain-service@1.1.13-hotfix.0
+  - @ledgerhq/evm-tools@1.0.8
+
 ## 6.34.7
 
 ### Patch Changes